refactor: move config loader to modules
